pyproject.toml
[tool.poetry]
name = "datasette-as-website"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.10"
datasette = "^0.64.1"
sqlite-utils = "^3.30"
datasette-render-markdown = "^2.1.1"
datasette-template-sql = "^1.0.2"
datasette-publish-fly = "^1.3"
click = "^8.1.3"
beautifulsoup4 = "^4.11.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
